add chat user bar invitation move
