+++ client-myf/src/app/components/chat/chat-right-sidebar/chat-user-bar-invitation/chat-user-bar-invitation.component.html
